requirements.txt
PyQt6>=6.6.0
pyqtgraph>=0.13.3
python-chess>=1.9.4
psutil>=5.9.0
nvidia-ml-py>=11.5.0
numpy>=1.24.0
torch>=2.0.0
numba>=0.58.0
